+++ azurerm/internal/services/containers/kubernetes_cluster_scaling_resource_test.go
@@ -290,7 +290,10 @@ func testAccKubernetesCluster_autoScalingProfile(t *testing.T) {
 				check.That(data.ResourceName).Key("default_node_pool.0.enable_auto_scaling").HasValue("true"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.expander").HasValue("least-waste"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.max_graceful_termination_sec").HasValue("15"),
+				check.That(data.ResourceName).Key("auto_scaler_profile.0.max_node_provisioning_time").HasValue("10m"),
+				check.That(data.ResourceName).Key("auto_scaler_profile.0.max_unready_percentage").HasValue("50"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.new_pod_scale_up_delay").HasValue("10s"),
+				check.That(data.ResourceName).Key("auto_scaler_profile.0.max_unready_nodes").HasValue("5"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.scale_down_delay_after_add").HasValue("10m"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.scale_down_delay_after_delete").HasValue("10s"),
 				check.That(data.ResourceName).Key("auto_scaler_profile.0.scale_down_delay_after_failure").HasValue("15m"),
@@ -551,6 +554,9 @@ resource "azurerm_kubernetes_cluster" "test" {
     balance_similar_node_groups      = true
     expander                         = "least-waste"
     max_graceful_termination_sec     = 15
+    max_node_provisioning_time       = "10m"
+    max_unready_nodes                = 5
+    max_unready_percentage           = 50
     new_pod_scale_up_delay           = "10s"
     scan_interval                    = "10s"
     scale_down_delay_after_add       = "10m"